#7 created a pip package and moved the existing examples and atoms there
