+++ python-runtime/scripts/build.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+cd ..
+python3 -m build
