+++ python-runtime/scripts/setup_env.sh
@@ -1,0 +1,8 @@
+#!/usr/bin/env bash
+# before running this, make sure that your venv is activated
+
+python3 -m pip install --upgrade build
+python3 -m pip install --upgrade twine
+python3 -m pip install --upgrade bump2version
+python3 -m pip install --upgrade eo2py
+
