+++ python-runtime/setup.cfg
@@ -1,0 +1,27 @@
+[metadata]
+name = eo2py
+version = 0.0.4
+author = Mikhail Olokin
+author_email = olomishcak@outlook.com
+description = Everything required to run Python files created by eo2py-maven-plugin
+license_files = LICENSE
+license = MIT
+long_description = file: README.md
+long_description_content_type = text/markdown
+url = https://github.com/nikololiahim/eo-runtime-python
+project_urls =
+    Bug Tracker = https://github.com/nikololiahim/eo-runtime-python/issues
+classifiers =
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Operating System :: OS Independent
+
+[options]
+package_dir =
+    = src
+packages = find:
+python_requires = >=3.8
+
+[options.packages.find]
+where = src
+
