+++ python-runtime/tests/factorial.py
@@ -1,4 +1,4 @@
-from atoms import *
+from eo2py.atoms import *
 
 """
 +package sandbox
